requirements.txt
asgiref
crispy-bootstrap5
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
Markdown
pillow
psycopg2
psycopg2-binary
PyJWT
pytz
sqlparse
typing_extensions
tzdata